web/requirements.txt
flask>=3.0.0
flask-socketio>=5.3.0
python-engineio>=4.8.0
python-socketio>=5.10.0
pymongo>=4.6.0
python-dotenv>=1.0.0
requests>=2.31.0
eventlet>=0.35.0
